agents/summarizer_agent/requirements.txt
flask
google-cloud-aiplatform==1.43.0